# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API permettant la gestion et la réservation de produits sous différentes formes, via des menus, des commandes. La mise en avant des produits ainsi que des promotions sont possible via celle-ci.",
    "version": "1.0.0",
    "title": "Burger API - Groupe 2",
    "license": {
      "name": "NodeJS",
      "url": "https://github.com/nodejs/node/blob/master/LICENSE"
    }
  },
  "host": "vincentguenin.ddns.net:8000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/product": {
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Ajouter un nouveau produit",
        "description": "",
        "operationId": "addProduct",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "description": "Token d'accès de l'utilisateur",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Produit à ajouter en base",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Succès de l'opération",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/modifyName/{oldName}/{newName}": {
      "patch": {
        "tags": [
          "product"
        ],
        "summary": "Modifier le nom d'un produit",
        "description": "",
        "operationId": "ModifyProductName",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "description": "Token d'accès",
            "required": true
          },
          {
            "in": "path",
            "name": "oldName",
            "type": "string",
            "description": "Ancien nom du produit",
            "required": true
          },
          {
            "in": "path",
            "name": "newName",
            "type": "string",
            "description": "Nouveau nom du produit",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Succès de l'opération"
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/modifyCal/{name}/{cal}": {
      "patch": {
        "tags": [
          "product"
        ],
        "summary": "Modifier le nombre de calories d'un produit",
        "description": "",
        "operationId": "ModifyProductCal",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "description": "Token d'accès",
            "required": true
          },
          {
            "in": "path",
            "name": "name",
            "type": "string",
            "description": "Nom du produit",
            "required": true
          },
          {
            "in": "path",
            "name": "cal",
            "type": "string",
            "description": "Nombre de calories d'un produit",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Succès de l'opération"
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/modifySize/{name}/{size}": {
      "patch": {
        "tags": [
          "product"
        ],
        "summary": "Modifier la taille d'un produit",
        "description": "",
        "operationId": "ModifyProductSize",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "description": "Token d'accès",
            "required": true
          },
          {
            "in": "path",
            "name": "name",
            "type": "string",
            "description": "Nom du produit",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "type": "string",
            "description": "Taille d'un produit",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Succès de l'opération"
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/modifyPrice/{name}/{price}": {
      "patch": {
        "tags": [
          "product"
        ],
        "summary": "Modifier le prix d'un produit",
        "description": "",
        "operationId": "ModifyProductPrice",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "description": "Token d'accès",
            "required": true
          },
          {
            "in": "path",
            "name": "name",
            "type": "string",
            "description": "Nom du produit",
            "required": true
          },
          {
            "in": "path",
            "name": "price",
            "type": "string",
            "description": "Prix d'un produit",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Succès de l'opération"
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/modifyHighlight/{id}": {
      "patch": {
        "tags": [
          "product"
        ],
        "summary": "Mettre en avant un produit",
        "description": "",
        "operationId": "ModifyProductHighlight",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "description": "Token d'accès",
            "required": true
          },
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "id du produit",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Succès de l'opération"
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/resetHighlight/{id}": {
      "patch": {
        "tags": [
          "product"
        ],
        "summary": "Retirer la mise en avant d'un produit",
        "description": "",
        "operationId": "resetProductHighlight",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "description": "Token d'accès",
            "required": true
          },
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "id du produit",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Succès de l'opération"
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/display/allProducts": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Affiche tous les produits",
        "description": "",
        "operationId": "deleteProduct",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Succès de l'opération",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/displayByName/{name}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Afficher un produit via son nom",
        "description": "",
        "operationId": "displayProductByName",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "type": "string",
            "description": "Nom du produit",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Succès de l'opération",
            "schema": {
              "$ref": "#/descriptions/Product"
            }
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/displayByPrice/{price}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Afficher tous les produits correspondant à un prix",
        "description": "",
        "operationId": "displayProductsByPrice",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "price",
            "type": "string",
            "description": "Prix du produit",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Succès de l'opération",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/descriptions/Product"
              }
            }
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/displayBySize/{size}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Afficher tous les produits correspondant à une taille",
        "description": "",
        "operationId": "displayProductsBySize",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "size",
            "type": "string",
            "description": "Taille du produit",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Succès de l'opération",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/descriptions/Product"
              }
            }
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/displayHighlight": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Afficher tous les produits mis en avant",
        "description": "",
        "operationId": "displayHighlightedProducts",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Succès de l'opération",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/descriptions/Product"
              }
            }
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
    "/product/delete/{id}": {
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Supprimer un produit",
        "description": "",
        "operationId": "deleteProduct",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "description": "Token d'accès",
            "required": true
          },
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "id du produit",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Succès de l'opération"
          },
          "400": {
            "description": "Paramètres invalides"
          },
          "500": {
            "description": "Erreur technique du serveur"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "cal": {
          "type": "integer",
          "format": "int64",
          "description": "Nombre de calories du produit"
        },
        "highlight": {
          "type": "boolean",
          "default": false
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "size": {
          "type": "string",
          "maxLength": 255
        }
      }
    },
...
  }
}